# Add Zstandard compression option for wlog