@leizor 2022-12-12
Oop, thanks for pointing that out @roidelapluie! I think I'll leave it for now as I'm not sure it's clear-cut whether it's desirable to include zstd as an option even.

If we do decide we'd like to, I'll revisit and get that fixed up.