@codesome 2022-12-14
How does WAL replay compare with this compression?