@SuperQ 2022-12-14
I agree, we need to keep the flags backwards compatible. But I think it's great to add bettter compression options.

The most important improvements to me are on-disk size, and startup speed reduction.