@leizor 2022-12-15
Yay, cool to see that there's some interest in this!

As far as the flag change goes, I'm thinking I'll go back and revert the `--compress` flag to what it used to do, but add a new `--compress-type` flag that can be either `snappy` or `zstd`.

I'll also work on grabbing some benchmark numbers for WAL replay (is this also what plays into startup speed reduction?).