@SuperQ 2022-12-15
Yup, WAL replay speed would be interesting to compare. We chose snappy originally due to the decompress speed.