@SuperQ 2022-12-20
Yes, ztsd is slower, more CPU heavy than snappy. Snappy was originally chosen to provide reasonable compression with minimal overhead.

One interesting option is adding to our snappy implementation with [klauspost S2](https://github.com/klauspost/compress/tree/master/s2#s2-compression).