@leizor 2022-12-20
> One interesting option is adding to our snappy implementation with [klauspost S2](https://github.com/klauspost/compress/tree/master/s2#s2-compression).

I did throw together a benchmark for S2 after I was disappointed with how zstd performed. I might've done something wrong though as it turned out to be slower than snappy still. I'll add it back in so we can double-check though!