@SuperQ 2022-12-21
I'm not surprised that zstd is slower, and that's fine. It's still a good option to have for users that are willing to trade off CPU / speed for better compression.

I think this feature is worth having, but maybe we stick with snappy as the default for now until we get some more experience with zstd.