@leizor 2022-12-22
Alright, in that case, sounds good! I think I'll just rollback that commit setting zstd as default and resolve the merge conflicts then, and this PR can be considered ready for review.

I think that way we can address the other compression options in follow-up PRs?