@SuperQ 2023-07-04
I would like to see support for zstd, higher compression ratios with only a bit more overhead.