@SuperQ 2023-07-11
I would love to see benchmark comparisons of `klauspost/compress` next.