@Whisper40 2025-04-25
Hello, how to use the walCompressionType parameter ?

When using Prometheus CR, this parameter cannot be used because the CRD has not this field.